+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 
 # 📑 Changelog
 
+## Added
+- [core] feature: added `command suggestions` for `greedy command string`, which benefits the following commands: `/command-cooldown test`, `/command-attachment {attach-entity|attach-item|attach-block} ...`, `/command-permission describe`, `/command_debug`
+
 ## 🐛 Fixed
 - **[core]**
     - The `--silent` and `--stdout` optional arguments can't be specified for `greedy command string` argument type.